+++ src/StatusBarUI.ts
@@ -1,42 +1,45 @@
 import * as vscode from "vscode";
 import { startServerID, stopServerID } from "./enum";
-import { Utils } from "./utils";
+import { Settings } from "./Settings";
 
 export class StatusbarUi {
-  statusBarItem!: vscode.StatusBarItem;
+  private static _statusBarItem: vscode.StatusBarItem;
 
-  constructor(showOnStatusbar: boolean) {
-    this.statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 100);
-    showOnStatusbar && this.statusBarItem.show();
-    this.Init();
+  static get statusBarItem() {
+    if (!StatusbarUi._statusBarItem) {
+      StatusbarUi._statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 100);
+      if (Settings.showOnStatusbar) this.statusBarItem.show();
+    }
+    return StatusbarUi._statusBarItem;
   }
 
-  Init() {
-    this.Working("loading...");
-    this.startServer(1000);
+  static Init() {
+    StatusbarUi.Working("loading...");
+    StatusbarUi.startServer(1000);
   }
 
-  Working(workingMsg = "Working on it...") {
-    this.statusBarItem.text = `$(pulse) ${workingMsg}`;
-    this.statusBarItem.tooltip = "In case if it takes long time, try to close all browser window.";
-    this.statusBarItem.command = undefined;
+  static Working(workingMsg = "Working on it...") {
+    StatusbarUi.statusBarItem.text = `$(pulse) ${workingMsg}`;
+    StatusbarUi.statusBarItem.tooltip = "In case if it takes long time, try to close all browser window.";
+    StatusbarUi.statusBarItem.command = undefined;
   }
 
-  startServer(delay: number, statusMsg?: string) {
+  static startServer(delay: number, showPopupMessage?: () => void) {
     setTimeout(() => {
-      this.statusBarItem.text = "$(broadcast) Mock it";
-      this.statusBarItem.command = startServerID;
-      this.statusBarItem.tooltip = "Click to start mock server";
-      statusMsg && vscode.window.showInformationMessage(statusMsg);
+      StatusbarUi.statusBarItem.text = "$(broadcast) Mock it";
+      StatusbarUi.statusBarItem.command = startServerID;
+      StatusbarUi.statusBarItem.tooltip = "Click to start mock server";
+      showPopupMessage && showPopupMessage();
     }, delay);
   }
 
-  stopServer(delay: number, port: number, statusMsg?: string) {
+  static stopServer(delay: number, port: number, showPopupMessage?: () => void) {
     setTimeout(() => {
-      this.statusBarItem.text = `$(circle-slash) Port : ${port}`;
-      this.statusBarItem.command = stopServerID;
-      this.statusBarItem.tooltip = "Click to stop mock server";
-      statusMsg && vscode.window.showInformationMessage(statusMsg);
+      StatusbarUi.statusBarItem.text = `$(circle-slash) Port : ${port}`;
+      StatusbarUi.statusBarItem.command = stopServerID;
+      StatusbarUi.statusBarItem.tooltip = "Click to stop mock server";
+      const dontshowTxt = "Don't show again";
+      showPopupMessage && showPopupMessage();
     }, delay);
   }
 }